deployment/requirements.txt
tensorflow
keras
numpy
pillow
gdown
requests
streamlit
plotly
